flask-auth-service/requirements.txt
flask
psycopg2
pyjwt
python-dotenv

